Use a foreach/2 that allows bail-out.
This commit allows us to bail out in the event that a single
transaction fails when issuing a bunch of deletes. I've also moved to
use DELETE rather than DELETEQ as the particulars of that operation
are poorly defined in documentation, is available only for the binary
protocol and I was uncomfortable trying to experimentally determine
its behaviour.

It surprised me and I got spooked.

Signed-off-by: Brian L. Troutwine <brian.troutwine@adroll.com>