Extend mero_SUITE to test the binary protocol.
This testing turned up a few bugs in the binary protocol, one around
the value reported back by Increment and another in the way status
codes get sent back to the client. There are still a few tests failing
that _only_ fail when run together in a group.

Fun times.

Signed-off-by: Brian L. Troutwine <brian.troutwine@adroll.com>