Minimum delay between connections
So far mero had support for throttling connections attempts when
there are connection failures.
This add a new, general throttling of connections (regardless if
they succed or not) as a min wait time between attempts.
It's useful when there are _large_ number of hosts connecting
to the same memcache backed.